First commit 03:23 26/10/2014
This is the first commit of the project Sangokushi a java based strategy
game featuring 550 three kingdoms characters, 30 different factions and
the Chinese landscape with 40 unique cities.

So far this projects has the following systems:

1. Allows players to select a Scenario at the beginning and then pick a
force of his or her own choice.

2. There are 2 different Scenarios, one is a fictional scenario and the
other one is real history based. In the fictional scenario there are 30
total number of different forces.

2. A GamePanel that displays the map landscape and 40 different cities.
An option menu is displayed when the players clicks on his or her owned
cities.

3. the option menu allows the player to do military, political or other
tasks. So far the dispatch troops and build structure systems have their
prototypes done.

4. The dispatch troops system allows player to chose which city to
attack and form legions with commanders and soldiers. A prototype for
the battle system is also done for testing wars between legions.

5. The build structures system allows the player to go into a specific
city and manage it's structures. Players can build various different
structures on the available empty slots. There is an initial of 5
different structures are available.

6. A proceed button that ends the turn for the player and updates the
time and resources.