Update: New Battle Logics
1. Done an entire new battle logics system, implemented from the game
Risk (3 dice are rolled and compared each round), However it is not
fully implemented into the game yet, due to existing bugs awaiting to be
fixed.
2. This new system supports Siege or Field battle scenarios, which will
allow to add more mechanics into the game.
3.  Optimised the BattleScreen class a bit more, added more
polymorphism.
4. Added 2 more unit types YellowTurban and QingZhou Troops.
5.  Changed ArrayList to LinkedList structure in BattleScreen and
implemented iterator for operations like move or translate of the whole
LinkedList. This was pirmarily to make use of a more efficient data
structure, since elements will be removed all the time one by one.
6. New Battle systems will be fully functional in the next update.